로테이션 픽스 & 게임 디버깅

+++ Assets/01.Scripts/Utils/Define.cs
@@ -61,5 +61,5 @@ public class Define
     public const int ELEMENT_MARBLE_COUNT = 3;
     #endregion
 
-    public const int FIXED_FPS = 50;
+    public const float FIXED_FPS = 50;
 }
